+++ test/extensions/string_extensions_test.dart
@@ -3,91 +3,91 @@ import 'package:test/test.dart';
 
 void main() {
   group("StringExtensions", () {
-    group("isEmptyOrNull", () {
-      test('should return true for null strings', () {
+    group("isEmptyOrNull should return", () {
+      test('true for null values', () {
         String? value;
         expect(value.isEmptyOrNull, true);
       });
 
-      test('should return true for empty strings', () {
+      test('true for empty values', () {
         const value = '';
         expect(value.isEmptyOrNull, true);
       });
 
-      test('should return false for non-empty strings', () {
+      test('false for non-empty values', () {
         const value = 'an actual value';
         expect(value.isEmptyOrNull, false);
       });
     });
 
-    group("isNotEmptyAndNull", () {
-      test('should return false for null strings', () {
+    group("isNotEmptyAndNull should return", () {
+      test('false for null values', () {
         String? value;
         expect(value.isNotEmptyAndNull, false);
       });
 
-      test('should return false for empty strings', () {
+      test('false for empty values', () {
         const value = '';
         expect(value.isNotEmptyAndNull, false);
       });
 
-      test('should return true for non-empty strings', () {
+      test('true for non-empty values', () {
         const value = 'an actual value';
         expect(value.isNotEmptyAndNull, true);
       });
     });
 
-    group("or", () {
+    group("or should return", () {
       const defaultValue = 'the default value';
 
-      test('should return the default value for null strings', () {
+      test('the default value for null values', () {
         String? value;
         expect(value.or(defaultValue), defaultValue);
       });
 
-      test('should return the default value for empty strings', () {
+      test('the default value for empty values', () {
         const value = '';
         expect(value.or(defaultValue), defaultValue);
       });
 
-      test('should return the actual value for non-empty strings', () {
+      test('the actual value for non-empty values', () {
         const value = 'the actual value';
         expect(value.or(defaultValue), value);
       });
     });
 
-    group("orEmpty", () {
-      test('should return an empty string for null strings', () {
+    group("orEmpty should return", () {
+      test('an empty string for null values', () {
         String? value;
         expect(value.orEmpty(), '');
       });
 
-      test('should return the same value for empty strings', () {
+      test('the same value for empty values', () {
         const value = '';
         expect(value.orEmpty(), value);
       });
 
-      test('should return the actual value for non-empty strings', () {
+      test('the actual value for non-empty values', () {
         const value = 'the actual value';
         expect(value.orEmpty(), value);
       });
     });
 
-    group("orProvider", () {
+    group("orProvider should return", () {
       const defaultValue = 'the default value';
       String defaultValueProvider() => defaultValue;
 
-      test('should return the default value for null strings', () {
+      test('the default value for null values', () {
         String? value;
         expect(value.orProvider(defaultValueProvider), defaultValue);
       });
 
-      test('should return the default value for empty strings', () {
+      test('the default value for empty values', () {
         const value = '';
         expect(value.orProvider(defaultValueProvider), defaultValue);
       });
 
-      test('should return the actual value for non-empty strings', () {
+      test('the actual value for non-empty values', () {
         const value = 'the actual value';
         expect(value.orProvider(defaultValueProvider), value);
       });