+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Olive's change logs
+
 ## 0.0.1
 
 * The very first version includes isEmptyOrNull, isNotEmptyAndNull, or, orEmpty, and orProvider extension methods for the String data type.
+
+## 0.0.2
+
+* This release includes limitMin, limitMax, and limitWithin for both int and double data types.
+* Also includes microseconds, milliseconds, seconds, minutes, hours, and days for the int data type.